refactor(client): extract xclient discovery into xclient_discovery.go
Move service-discovery, server-filtering, and cached-client-management
methods out of xclient.go (God File, Critical #1) into a focused
xclient_discovery.go. Same package, pure mechanical move — no logic,
API, or behavior change.

Moved: watch, filterByStateAndGroup, selectClient, safeCloseClient,
getCachedClient/setCachedClient/findCachedClient/deleteCachedClient/
removeClient/generateClient/getCachedClientWithoutLock,
splitNetworkAndAddress.
